+++ app/routes/__init__.py
@@ -4,7 +4,6 @@ from flask_restx import Api
 from app.web import register_namespaces as register_web_namespaces
 
 from .account_route import account_bp
-from .namespaces.login_ns import login_ns
 from .namespaces.role_ns import role_ns
 from .namespaces.subscrtiption_ns import subscription_ns
 
@@ -15,7 +14,6 @@ def register_routes(app: Flask):
 
 
 def register_namespaces(api: Api):
-    api.add_namespace(login_ns)
     api.add_namespace(role_ns)
     api.add_namespace(subscription_ns)
     register_web_namespaces(api)