+++ app/web/__init__.py
@@ -1,5 +1,6 @@
 from flask_restx import Api
 
+from .auth.routes import login_ns
 from .leaves.routes import leave_ns
 from .api_keys.routes import api_key_ns
 from .cities.routes import city_ns
@@ -21,6 +22,7 @@ from .works.routes import work_ns
 
 
 def register_namespaces(api: Api):
+    api.add_namespace(login_ns)
     api.add_namespace(api_key_ns)
     api.add_namespace(leave_ns)
     api.add_namespace(city_ns)